refactor(docs): date-range-picker dx
