+++ apps/docs/content/components/date-range-picker/description.raw.jsx
@@ -1,0 +1,11 @@
+import {DateRangePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <DateRangePicker
+      className="max-w-xs"
+      description="Please enter your stay duration"
+      label="Stay duration"
+    />
+  );
+}
