+++ apps/docs/content/components/date-range-picker/description.ts
@@ -1,14 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <DateRangePicker
-      label="Stay duration"
-      description="Please enter your stay duration"
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./description.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
