+++ apps/docs/content/components/date-range-picker/error-message-function.ts
@@ -1,25 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-import {parseDate} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <DateRangePicker 
-      isInvalid
-      label="Stay duration" 
-      variant="bordered"
-      errorMessage={(value) => {
-        if (value.isInvalid) {
-          return "Please enter your stay duration";
-        }
-      }}
-      defaultValue={{
-        start: parseDate("2024-04-01"),
-        end: parseDate("2024-04-08"),
-      }}
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./error-message-function.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
