+++ apps/docs/content/components/date-range-picker/label-placements.ts
@@ -1,27 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-
-export default function App() {
-  const placements = [
-    "inside",
-    "outside",
-    "outside-left",
-  ];
-
-  return (
-    <div className="flex flex-col gap-4">
-      <div className="flex flex-col gap-2">
-        {placements.map((placement) => (
-          <DateRangePicker
-            key={placement}
-            label="Stay duration"
-            labelPlacement={placement}
-            description={placement}
-          />
-        ))}
-      </div>  
-    </div>  
-  );
-}`;
+import App from "./label-placements.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
