+++ apps/docs/content/components/date-range-picker/non-contiguous.ts
@@ -1,20 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-import {isWeekend, today, getLocalTimeZone} from "@internationalized/date";
-import {useLocale} from "@react-aria/i18n";
-
-export default function App() {
-  let {locale} = useLocale();
-
-  return (
-    <DateRangePicker
-      allowsNonContiguousRanges
-      isDateUnavailable={(date) => isWeekend(date, locale)}
-      label="Time off request"
-      minValue={today(getLocalTimeZone())}
-      visibleMonths={2}
-    />
-  );
-}`;
+import App from "./non-contiguous.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
