+++ apps/docs/content/components/date-range-picker/page-behavior.ts
@@ -1,16 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <DateRangePicker 
-        label="Birth date"
-        visibleMonths={2}
-        pageBehavior="single"
-      />
-    </div> 
-  );
-}`;
+import App from "./page-behavior.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
