+++ apps/docs/content/components/date-range-picker/readonly.ts
@@ -1,19 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-import {parseDate} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <DateRangePicker 
-      label="Stay duration" 
-      isReadOnly
-      defaultValue={{
-        start: parseDate("2024-04-01"),
-        end: parseDate("2024-04-08"),
-      }}
-      className="max-w-xs"
-    />
-  );
-}`;
+import App from "./readonly.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
