+++ apps/docs/content/components/date-range-picker/usage.raw.jsx
@@ -1,0 +1,5 @@
+import {DateRangePicker} from "@nextui-org/react";
+
+export default function App() {
+  return <DateRangePicker className="max-w-xs" label="Stay duration" />;
+}
