+++ apps/docs/content/components/date-range-picker/visible-month.raw.jsx
@@ -1,0 +1,9 @@
+import {DateRangePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
+      <DateRangePicker label="Stay duration" visibleMonths={2} />
+    </div>
+  );
+}
