+++ apps/docs/content/components/date-range-picker/visible-month.ts
@@ -1,15 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <div className="flex w-full flex-wrap md:flex-nowrap gap-4">
-      <DateRangePicker 
-        label="Stay duration"
-        visibleMonths={2}
-      />
-    </div> 
-  );
-}`;
+import App from "./visible-month.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
