+++ apps/docs/content/components/date-range-picker/with-month-and-year-pickers.raw.jsx
@@ -1,0 +1,9 @@
+import {DateRangePicker} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <div className="w-full max-w-xl flex flex-row gap-4">
+      <DateRangePicker showMonthAndYearPickers label="Birth Date" variant="bordered" />
+    </div>
+  );
+}
