+++ apps/docs/content/components/date-range-picker/with-time-field.ts
@@ -1,21 +1,4 @@
-const App = `import {DateRangePicker} from "@nextui-org/react";
-import {parseZonedDateTime} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <div className="w-full max-w-xl flex flex-row gap-4">
-      <DateRangePicker
-        label="Event duration"
-        hideTimeZone
-        visibleMonths={2}
-        defaultValue={{
-          start: parseZonedDateTime("2024-04-01T00:45[America/Los_Angeles]"),
-          end: parseZonedDateTime("2024-04-08T11:15[America/Los_Angeles]"),
-        }}
-      />
-    </div>
-  );
-}`;
+import App from "./with-time-field.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
